requirements.txt
datetime
pyjokes
requests
Pillow
Image
ImageGrab
gTTS
keyboard
key
SpeechRecognition
